automation-project1/automation-project/requirements.txt
requests~=2.31.0
bs4~=0.0.2
beautifulsoup4~=4.12.3
matplotlib~=3.8.4
Pillow~=10.1.0
colorthief~=0.2.1